+++ src/error.rs
@@ -1,8 +1,8 @@
-#[cfg(feature = "varing")]
-#[cfg_attr(docsrs, doc(cfg(feature = "varing")))]
+#[cfg(feature = "varint")]
+#[cfg_attr(docsrs, doc(cfg(feature = "varint")))]
 pub use varing::InsufficientSpace;
 
-#[cfg(feature = "varing")]
+#[cfg(feature = "varint")]
 pub use varing::{
   ConstDecodeError as ConstDecodeVarintError, ConstEncodeError as ConstEncodeVarintError,
   DecodeError as DecodeVarintError, EncodeError as EncodeVarintError,
@@ -639,8 +639,8 @@ impl From<TryPutAtError> for std::io::Error {
 /// An error that occurs when trying to put type in LEB128 format at a specific offset in the buffer.
 #[derive(Debug, Clone, PartialEq, Eq, thiserror::Error)]
 #[non_exhaustive]
-#[cfg(feature = "varing")]
-#[cfg_attr(docsrs, doc(cfg(feature = "varing")))]
+#[cfg(feature = "varint")]
+#[cfg_attr(docsrs, doc(cfg(feature = "varint")))]
 pub enum EncodeVarintAtError {
   /// The offset is out of bounds for the buffer length.
   #[error(transparent)]
@@ -659,7 +659,7 @@ pub enum EncodeVarintAtError {
   Other(std::borrow::Cow<'static, str>),
 }
 
-#[cfg(feature = "varing")]
+#[cfg(feature = "varint")]
 impl EncodeVarintAtError {
   /// Creates a new `EncodeVarintAtError::OutOfBounds` error.
   #[inline]
@@ -726,7 +726,7 @@ impl EncodeVarintAtError {
   }
 }
 
-#[cfg(all(feature = "varing", feature = "std"))]
+#[cfg(all(feature = "varint", feature = "std"))]
 impl From<EncodeVarintAtError> for std::io::Error {
   fn from(e: EncodeVarintAtError) -> Self {
     match e {
@@ -748,8 +748,8 @@ impl From<EncodeVarintAtError> for std::io::Error {
 /// An error that occurs when trying to put type in LEB128 format at a specific offset in the buffer.
 #[derive(Debug, Clone, PartialEq, Eq, thiserror::Error)]
 #[non_exhaustive]
-#[cfg(feature = "varing")]
-#[cfg_attr(docsrs, doc(cfg(feature = "varing")))]
+#[cfg(feature = "varint")]
+#[cfg_attr(docsrs, doc(cfg(feature = "varint")))]
 pub enum DecodeVarintAtError {
   /// The offset is out of bounds for the buffer length.
   #[error(transparent)]
@@ -768,7 +768,7 @@ pub enum DecodeVarintAtError {
   Other(std::borrow::Cow<'static, str>),
 }
 
-#[cfg(feature = "varing")]
+#[cfg(feature = "varint")]
 impl DecodeVarintAtError {
   /// Creates a new `DecodeVarintAtError::OutOfBounds` error.
   #[inline]
@@ -832,7 +832,7 @@ impl DecodeVarintAtError {
   }
 }
 
-#[cfg(all(feature = "varing", feature = "std"))]
+#[cfg(all(feature = "varint", feature = "std"))]
 impl From<DecodeVarintAtError> for std::io::Error {
   fn from(e: DecodeVarintAtError) -> Self {
     match e {