+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add consume and scan varint related methods for `ChunkExt`
 - Change type of `requested` from `usize` to `NonZeroUsize`
 - Bumpup `varing` to `0.10`
+- Change feature `varing` to feature `varint`
 
 ## 0.4.0 (Aug 11st, 2025)